requirements.txt
Django==3.1.1
django-crispy-forms==1.9.2
numpy==1.19.1
pandas==1.1.1
django-pandas==0.6.2
gspread-pandas==2.2.3
djangorestframework==3.11.1
Markdown==3.2.2
djangorestframework-jwt==1.11.0